requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.29
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.3
uritemplate==4.1.1